requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
chromadb
sentence-transformers
streamlit
pandas
pydantic
pydantic-settings
python-dotenv
google-generativeai
neo4j
